Splitting file and changing test back
