+++ tests/behaviour/contracts/memory/dynamicArrays.sol
@@ -4,10 +4,8 @@ pragma solidity ^0.8.6;
 
 contract WARP {
     function uint8new() pure public returns (uint8, uint8) {
-        uint8[] memory x = new uint8[](5);
-        x[0] = 5;
-        x[1] = 1;
-        return (0, 1);
+        uint8[] memory x = new uint8[](2);
+        return (x[0], x[1]);
     }
 
     function uint8write(uint8 v) pure public returns (uint8, uint8) {
